robot/requirements.txt
beautifulsoup4
requests
psycopg2-binary
pyyaml
lxml